Revert reveal.js back to unforked version.
For links into and within presentations, reveal.js understands named
links as documented there:
https://github.com/hakimel/reveal.js#internal-links

However, for a slide with named id “target” those links are of the
form “#/target”, which is an invalid fragment identifier.  Slashes are
allowed in URLs, but then they represent data.  I would prefer valid
links.  My attempt to fix this is documented in this issue (if you are
reading this, maybe you can help?):
https://github.com/hakimel/reveal.js/issues/2276

As stated in that issue, when using valid URLs, I need to choose
between broken history functionality and jumping contents upon slide
transitions.  For the time being, I choose broken history
functionality, which I perceive to be the minor annoyance.